Replace fontdb with fc-match, add a11y tree, harden runtime
Major changes:
- Drop fontdb dependency in favor of fc-match/fc-list CLI for font lookup,
  eliminating the ~200ms startup scan of all system fonts
- Add accessibility tree scaffolding (A11yNode/A11yRole) with node emission
  from all interactive widgets, high-contrast theme, gsettings detection
- Widen glyph_id from u16 to u32 to match rustybuzz's native type
- Convert scene graph traversal/removal from recursive to iterative
  to prevent stack overflow on deep widget trees
- Strip wgpu to vulkan+wgsl only, add release profile LTO/strip/codegen-units=1
- Add GPU device-lost recovery (exit after 3 consecutive render failures)
- Use sigaction instead of signal() for proper SIGTERM/SIGINT handling
- Prune stale scroll offsets after ~5s of inactivity
- Make TextRenderer::new() fallible instead of panicking
- Fix atlas rebind to check bound flag rather than upload presence

+++ crates/esox_font/examples/visual_demo.rs
@@ -84,7 +84,7 @@ fn main() {
         println!("--- Cell clipping analysis ---");
         for ch in text.chars() {
             let font_ref = face.as_swash_ref();
-            let glyph_id = font_ref.charmap().map(ch);
+            let glyph_id = u32::from(font_ref.charmap().map(ch));
             let rast = rasterizer.rasterize(&face, glyph_id, size, 0).unwrap();
             let glyph_top = metrics.ascent - rast.bearing_y;
             let glyph_bottom = glyph_top + rast.height as f32;
